Move tests into each util module.
