+++ src/util/mod.rs
@@ -17,33 +17,23 @@ mod util_macos;
 // Re-export acl_entry_t and acl_t from crate::sys.
 pub use crate::sys::{acl_entry_t, acl_t};
 
-pub use util_common::{
-    xacl_create_entry, xacl_entry_count, xacl_foreach, xacl_free, xacl_from_text, xacl_init,
-    xacl_is_empty, xacl_to_text,
-};
-
 #[cfg(target_os = "freebsd")]
 pub use util_freebsd::{
-    xacl_add_entry, xacl_get_entry, xacl_get_file, xacl_is_posix, xacl_set_file,
+    xacl_add_entry, xacl_create_entry, xacl_entry_count, xacl_foreach, xacl_free, xacl_from_text,
+    xacl_get_entry, xacl_get_file, xacl_init, xacl_is_empty, xacl_is_posix, xacl_set_file,
+    xacl_to_text,
 };
 
 #[cfg(target_os = "linux")]
-pub use util_linux::{xacl_add_entry, xacl_get_entry, xacl_get_file, xacl_is_posix, xacl_set_file};
+pub use util_linux::{
+    xacl_add_entry, xacl_create_entry, xacl_entry_count, xacl_foreach, xacl_free, xacl_from_text,
+    xacl_get_entry, xacl_get_file, xacl_init, xacl_is_empty, xacl_is_posix, xacl_set_file,
+    xacl_to_text,
+};
 
 #[cfg(target_os = "macos")]
 pub use util_macos::{
-    xacl_add_entry, xacl_get_acl_flags, xacl_get_entry, xacl_get_file, xacl_is_posix,
-    xacl_set_acl_flags, xacl_set_file,
+    xacl_add_entry, xacl_create_entry, xacl_entry_count, xacl_foreach, xacl_free, xacl_from_text,
+    xacl_get_acl_flags, xacl_get_entry, xacl_get_file, xacl_init, xacl_is_empty, xacl_is_posix,
+    xacl_set_acl_flags, xacl_set_file, xacl_to_text,
 };
-
-#[cfg(test)]
-#[cfg(target_os = "freebsd")]
-mod util_freebsd_test;
-
-#[cfg(test)]
-#[cfg(target_os = "linux")]
-mod util_linux_test;
-
-#[cfg(test)]
-#[cfg(target_os = "macos")]
-mod util_macos_test;
